#22 added tests for docs

+++ BurnSoft.Applications.MGC/Firearms/Documents.cs
@@ -199,7 +199,7 @@ namespace BurnSoft.Applications.MGC.Firearms
         /// <returns><c>true</c> if XXXX, <c>false</c> otherwise.</returns>
         /// <exception cref="System.Exception"></exception>
         /// <exception cref="System.Exception"></exception>
-        public static bool Delete(string databasePath, int id, out string errOut)
+        public static bool Delete(string databasePath, long id, out string errOut)
         {
             bool bAns = false;
             errOut = @"";
